fix: remove unnecessary subscription
The query and subscription streams use the Wonka share operator to allow
multiple subscriptions. Calling publish on this stream attaches an
additional sink to the share operator that can not be removed. Because
of this the stream will never end and on every component mount another
query/subscription will be created.

I don't see any need for this call and everything works as advertised
without it.